Vision.jl

Vision.jl

Functions

Index

Documentation for Vision.jl

Functions

Vision.makeRequestBody — Function

makeRequestBody(image, features)

Create the request body for the Google Vision API.

Arguments

image base64 encoded image of type String or URI of type URI

features::Array{Dict{String, Any}} list of features to request from the API

Returns

Dict{String, Any} request body

Vision.visionFeature — Function

makeRequest(featureType::String, maxResults::Int=10)

Create a dictionary containing the feature type and max results.

Arguments

featureType::String type of feature to request from the API

maxResults::Int=10 maximum number of results to return

Vision.getResponse — Function

getResponse(requestBody, URL, headers)

Make a request to the Google Vision API and return as a dictionary.

Arguments

requestBody JSON request body

URL URL of the API, defaults to https://vision.googleapis.com/v1/images:annotate, this can be changed to use a different API or if it isn't possible to load environment variables.

headers headers for the request, defaults to []

Returns

Dict{String, Any} response from the API

Vision.parseFeatures — Method

parseFeatures(responseBody)

Parse the response body from the Google Vision API, returns the dictionary if the method hasn't been implemented. Otherwise return formatted output.

Arguments

responseBody response body from the API

Returns

Dictionary containing raw features or formatted output

Index

Vision.getResponse

Vision.makeRequestBody

Vision.parseFeatures

Vision.visionFeature